Introduce the {year} wildcard for multiple weather year support
The {year} wildcard represents a single year or a set of years, for
example "2013" or "1980-2000". The specification also supports
non-contiguous sets of years, for example
"2000-2002+2010" (representing the set of year [2000, 2001, 2002,
2010]).

This means that PyPSA-Eur functionality has been extended in two ways:
1. Networks can be defined over different weather years,
2. Networks can be defined over periods longer than a single years.

One the other hand, with this commit PyPSA-Eur is limited to networks
defined over some whole number of years; networks defined over smaller
periods than an entire year are no longer supported. Specifically, the
`snapshots.{start, end}` config is now ignored, and instead the
`snapshots.year_boundary` options specifies on which date model years
are to start.

It is now necessary for the user to download the required cutouts and
place them in the "cutouts" directory.

Implementation-wise, the main changes in this commit involve:
- Updating filenames in the Snakefile to include the {year} wildcard.
- Setting the network snapshots only in `add_electricity.py`, using
  the `year_boundary` config and {year} wildcard.
- Selecting the required cutouts (in the Snakefile) and stitching them
  together into renewable profiles (in `build_renewable_profiles.py`
  and `build_hydro_profile.py`)
- Updating how the number of years over which the network is defined
  is calculated (for cost calculations).