Fix reading custom generator attr files specified in config
These files are listed in the pypsa-eur config and given as inputs to
the `add_electricity' rule. However, when run as a submodule, the
filenames may come with a prefix, which is correctly handled by the
snakemake rule but not present in the specification in the config.
This commit updates the code reading the file to use the filename
supplied by snakemake, which comes with the correct prefix.