Remove 2013 cutout configuration from atlite
We don't want this configuration because it would cause the respective
cutouts to be downloaded and used in building natura and ship
rasters. (The rules building these rasters use _every_ cutout that has
configuration as input, which is a problem in the rule design.)
We want to use pypsa-eur as a snakemake module and update relevant
parts of the default configuration, but it's extra work to _remove_
parts of the default configuration later on. So we just remove these
sections at the source.